move ErrorHandler
